Struct zim::Zim

pub struct Zim {
    pub article_count: u32,
    pub cluster_count: u32,
    pub main_page_idx: Option<u32>,
    // some fields omitted
}

Represents a ZIM file

Fields

article_count

Number of articles in this archive

cluster_count

Number of clusters in this archive

main_page_idx

If Main Page is defined, this is the index to the page

Methods

impl Zim

fn new<P: AsRef<Path>>(p: P) -> Result<Zim, ParsingError>

Loads a Zim file

Loads a Zim file and parses the header, and the url, title, and cluster offset tables. The rest of the data isn't parsed until it's needed, so this should be fairly quick.

fn get_mimetype(&self, id: u16) -> Option<MimeType>

Indexes into the ZIM mime_table.

fn iterate_by_urls(&self) -> DirectoryIterator

Iterates over articles, sorted by URL.

For performance reasons, you might want to extract by cluster instead.

fn get_by_url_index(&self, idx: u32) -> Option<DirectoryEntry>

Returns the DirectoryEntry for the article found at the given URL index.

idx must be between 0 and article_count

fn get_cluster(&self, idx: u32) -> Option<Cluster>

Returns the given Cluster

idx must be between 0 and cluster_count